Reverted change that made DisplayHost's scale int-based.  It was a failed attempt to sidestep pixel rounding errors, and since it didn't work, we might as well be able to scle the display at a finer grained level
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@765041 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/pivot/wtk/Display.java
@@ -81,7 +81,7 @@ public final class Display extends Container {
         if (immediate) {
             Graphics2D graphics = (Graphics2D)displayHost.getGraphics();
 
-            int scale = displayHost.getScale();
+            double scale = displayHost.getScale();
             if (scale != 1) {
                 graphics.scale(scale, scale);
             }
